+++ conans/conan.py
@@ -2,7 +2,7 @@ import sys
 import os
 
 if os.getenv("CONAN_V2_CLI"):
-    from conans.cli.command import main
+    from conans.cli.cli import main
 else:
     from conans.client.command import main
